requirements.txt
streamlit
Pillow
python-dotenv
generativeai